Improve stall detection and resolution (#1808)
